limit and file removeal
