+++ lib/utils.js
@@ -1,3 +1,4 @@
+var extent = require('geojson-extent');
 var queue = require('queue-async');
 var geobuf = require('geobuf');
 var Pbf = require('pbf');
@@ -5,7 +6,6 @@ var url = require('url');
 var geojsonNormalize = require('geojson-normalize');
 var _ = require('lodash');
 var cuid = require('cuid');
-var Metadata = require('./metadata');
 
 module.exports = Utils;
 
@@ -69,9 +69,9 @@ function Utils(config) {
      */
 
     utils.createFeatureKey = function(dataset, featureId) {
-      return {
-        index: dataset + '!feature!' + featureId
-      }
+        return {
+            index: dataset + '!feature!' + featureId
+        }
     };
 
     utils.resolveFeaturesByIds = function(dataset, featureIds, callback) {
@@ -134,20 +134,27 @@ function Utils(config) {
             throw new Error('Unlocated features can not be stored.');
         }
 
-        var info = Metadata(null, null, null).getFeatureInfo(f);
+        var bounds = extent(f);
+        var info = {
+            bounds: bounds,
+            west: bounds[0],
+            south: bounds[1],
+            east: bounds[2],
+            north: bounds[3]
+        };
         var buf = utils.encodeFeature(f);
         var useS3 = buf.length >= config.MAX_GEOMETRY_SIZE;
         var s3Key = [config.prefix, dataset, primary, +new Date()].join('/');
         var s3Params = { Bucket: config.bucket, Key: s3Key, Body: buf };
 
         var databaseFeature = utils.createFeatureKey(dataset, primary);
-        databaseFeature.val = useS3 ? null : buf,
-        databaseFeature.size = info.size,
-        databaseFeature.west = truncateNum(info.west),
-        databaseFeature.south = truncateNum(info.south),
-        databaseFeature.east = truncateNum(info.east),
-        databaseFeature.north = truncateNum(info.north),
-        databaseFeature.s3url = useS3 ? ['s3:/', config.bucket, s3Key].join('/') : null
+        databaseFeature.val = useS3 ? null : buf;
+        databaseFeature.size = buf.length;
+        databaseFeature.west = truncateNum(info.west);
+        databaseFeature.south = truncateNum(info.south);
+        databaseFeature.east = truncateNum(info.east);
+        databaseFeature.north = truncateNum(info.north);
+        databaseFeature.s3url = useS3 ? ['s3:/', config.bucket, s3Key].join('/') : null;
 
         return {
             feature: databaseFeature,
